requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
azure-storage-blob
azure-keyvault-secrets
pandas
urllib3
requests 
azure-identity
sqlalchemy
pyodbc